可以跑通onnx图了 但是误差还很大

+++ sw/java/main/java/org/forwarder/backend/impls/HWAccelerated/opsets/v13/ops/HWAcceleratedGemmV13.java
@@ -49,7 +49,7 @@ public class HWAcceleratedGemmV13 extends HWAcceleratedOperator implements GemmV
 
         HWAcceleratedMatMulV13 matmulOp = new HWAcceleratedMatMulV13();
         INDArray matmulResult = matmulOp.matmul(A, B, fpgaInShift, fpgaOutShift);
-        INDArray Y = matmulResult.mul(alpha);
+        INDArray Y = matmulResult.mul((int)alpha);
 
         if (C != null) {
             long[] yShape = Y.shape();
